+++ packages/components/src/List/ListItem.story.tsx
@@ -1,0 +1,102 @@
+/*
+
+ MIT License
+
+ Copyright (c) 2020 Looker Data Sciences, Inc.
+
+ Permission is hereby granted, free of charge, to any person obtaining a copy
+ of this software and associated documentation files (the "Software"), to deal
+ in the Software without restriction, including without limitation the rights
+ to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ copies of the Software, and to permit persons to whom the Software is
+ furnished to do so, subject to the following conditions:
+
+ The above copyright notice and this permission notice shall be included in all
+ copies or substantial portions of the Software.
+
+ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ SOFTWARE.
+
+ */
+
+import React from 'react'
+import { Story } from '@storybook/react/types-6-0'
+
+import { ListItem, ListItemProps } from './ListItem'
+
+const Template: Story<ListItemProps> = (args) => (
+  <ListItem {...args}>Menu Item</ListItem>
+)
+
+export const Basic = Template.bind({})
+
+export const Icon = Template.bind({})
+Icon.args = {
+  ...Basic,
+  icon: 'User',
+}
+
+export const Detail = Template.bind({})
+Detail.args = {
+  ...Basic,
+  detail: 'A Detail',
+}
+
+export const IconAndDetail = Template.bind({})
+IconAndDetail.args = {
+  ...Basic,
+  detail: 'A Detail',
+  icon: 'User',
+}
+
+export const Description = Template.bind({})
+Description.args = {
+  ...Basic,
+  description: 'A description',
+}
+
+export const IconAndDescription = Template.bind({})
+IconAndDescription.args = {
+  ...Basic,
+  description: 'A description',
+  icon: 'User',
+}
+
+export const DetailAndDescription = Template.bind({})
+DetailAndDescription.args = {
+  ...Basic,
+  description: 'A description',
+  detail: 'A detail',
+}
+
+export const IconAndDetailAndDescription = Template.bind({})
+IconAndDetailAndDescription.args = {
+  ...Basic,
+  description: 'A description',
+  detail: 'A detail',
+  icon: 'User',
+}
+
+export const Current = Template.bind({})
+Current.args = {
+  ...Basic,
+  current: true,
+}
+
+export const Link = () => {
+  return (
+    <ListItem itemRole="link" href="https://google.com" target="_blank">
+      ListItem that links to Google
+    </ListItem>
+  )
+}
+
+export default {
+  component: ListItem,
+  title: 'ListItem',
+}