+++ packages/components/src/List/index.ts
@@ -24,11 +24,5 @@
 
  */
 
-import styled from 'styled-components'
-
-export const ListItemDetail = styled.div`
-  color: ${({ theme: { colors } }) => colors.text1};
-  font-size: ${({ theme: { fontSizes } }) => fontSizes.xsmall};
-  margin-left: auto;
-  padding-left: ${({ theme: { space } }) => space.xsmall};
-`
+export * from './List'
+export * from './ListItem'